@@ -8,19 +8,17 @@
     "deploy": "webpack -p --config webpack.config.production.js;"
   },
   "dependencies": {
-    "babel-core": "^5.8.22",
-    "babel-loader": "^5.3.2",
+    "babel-core": "^6.2.1",
+    "babel-loader": "^6.2.0",
+    "babel-preset-es2015": "^6.1.18",
+    "babel-preset-react": "^6.1.18",
     "body-parser": "~1.13.2",
     "cookie-parser": "~1.3.5",
-    "css-loader": "^0.16.0",
     "debug": "~2.2.0",
-    "ejs": "~2.3.3",
     "eslint-plugin-react": "^3.2.3",
     "express": "~4.13.1",
     "forever": "^0.15.1",
-    "grunt": "^0.4.5",
-    "grunt-webpack": "^1.0.8",
-    "history": "^1.9.0",
+    "history": "~1.13.0",
     "html-webpack-plugin": "^1.6.1",
     "immutable": "^3.7.4",
     "jquery": "^2.1.4",
@@ -30,16 +28,14 @@
     "morgan": "~1.6.1",
     "node-jsx": "^0.13.3",
     "node-libs-browser": "^0.5.2",
-    "node-sass": "^3.2.0",
     "q": "^1.4.1",
+    "radium": "^0.14.3",
     "ramda": "^0.17.1",
-    "react-doc-meta": "^0.2.0",
-    "react-geosuggest": "^1.8.0",
-    "react-loader": "^1.4.0",
-    "react-motion": "^0.3.0",
-    "react-router": "^1.0.0-rc1",
+    "react-copy-to-clipboard": "^3.0.4",
+    "react-geosuggest": "^1.13.0",
+    "react-loader": "^2.0.0",
+    "react-motion": "^0.3.1",
     "rx": "^3.1.0",
-    "sass-loader": "^2.0.1",
     "serve-favicon": "~2.3.0",
     "soundmanager2": "^2.97.20150601-a",
     "style-loader": "^0.12.3",
@@ -48,14 +44,10 @@
     "webpack-dev-server": "^1.10.1"
   },
   "devDependencies": {
-    "browserify": "^11.0.1",
-    "flux": "^2.0.3",
-    "gulp": "^3.9.0",
     "http-proxy": "^1.11.2",
-    "react": "^0.13.3",
+    "json-loader": "^0.5.3",
+    "react": "^0.14.3",
     "react-tools": "^0.13.3",
-    "reactify": "^1.1.1",
-    "vinyl-source-stream": "^1.1.0",
-    "watchify": "^3.3.1"
+    "vinyl-source-stream": "^1.1.0"
   }
 }
