@@ -3,23 +3,25 @@ import ArtistTile from './ArtistTile';
 
 var ArtistTileContainer = React.createClass({
 
-	render: function() {
-		var artists = this.props.artists;
-		var push = this.props.push;
+	getDefaultProps() {
+		return {
+			artists: []
+		};
+	},
 
-		var tiles = artists.map(function(artist) {
-			var props = {
+	render: function() {
+		var tiles = this.props.artists.map(artist => {
+			return React.createElement(ArtistTile, {
 				artist: artist.artist,
 				key: artist.id,
-				push: push,
-				imageURL: artist.imageURL
-			};
-
-			return <ArtistTile {...props} />;
+				imageURL: artist.icon_image.imageURL,
+				set_count: artist.set_count,
+				event_count: artist.event_count
+			})
 		});
 
 		return (
-			<div className='results-container flex-row flex'>
+			<div className='flex-row flex'>
 				{tiles}
 			</div>
 		);
@@ -27,4 +29,4 @@ var ArtistTileContainer = React.createClass({
 
 });
 
-module.exports = ArtistTileContainer;
+export default ArtistTileContainer;
