@@ -1,44 +1,37 @@
 import React from 'react';
 import Loader from 'react-loader';
 
-import FeaturedContainer from './FeaturedContainer';
-import LocationModule from './LocationModule';
+import FeaturedEvents from './FeaturedEvents';
+import Location from './Location';
 import EventContainer from './EventContainer';
 
-var EventsView = React.createClass({
-
-	componentDidMount: function() {
+const EventsView = React.createClass({
+	displayName: 'Events Page',
+	componentDidMount() {
 		mixpanel.track("Events Page Open");
 	},
-	
-	displayName: 'EventsView',
-	render: function() {
-		var push = this.props.push;
-		var appState = this.props.appState;
-		var containerClass='flex-row results-container tile-container';
+	render() {
+		var {appState, push} = this.props;
 
 		return (
-			<div id="EventsView" className="view flex-column">
+			<div id='EventsView' className='view flex-column'>
 				<div className='view-title-container flex-column'>
 					<h3 className='center'>Featured</h3>
 					<div className='divider'/>
 				</div>
-				<FeaturedContainer 
+				<FeaturedEvents 
 					appState={appState}
 					push={push} />
-				<LocationModule
-					push={push}
-                	appState={appState} />
-            	{
+				<Location appState={appState} />
+	        	{
 					React.cloneElement(this.props.children, {
 						appState: appState,
-						push: push,
-						containerClass: containerClass
+						push: push
 					})
 				}
-          </div>
+			</div>
 		);
 	}
 });
 
-module.exports = EventsView
+export default EventsView
