@@ -1,95 +1,66 @@
 import React from 'react';
 import R from 'ramda';
-import constants from '../constants/constants';
+import api from '../services/api';
 import Loader from 'react-loader';
+import {DEFAULT_IMAGE} from '../constants/constants';
 
 import SetContainer from './SetContainer';
 import DetailImageContainer from './DetailImageContainer';
 
-var FestivalDetail = React.createClass({
-
-	displayName: 'FestivalDetail',
-	getInitialState: function() {
+const FestivalDetail = React.createClass({
+	getInitialState() {
 		return {
-			loaded: false
+			loaded: false,
+			festival: '',
+			sets: [],
+			imageURL: DEFAULT_IMAGE
 		};
 	},
 
-	componentWillMount: function() {
-		this.getFestivalData();
+	componentWillMount() {
+		this.getFestivalData(this.props.params.festival);
 	},
 
-	getFestivalData: function() {
-		var _this = this;
-		var push = this.props.push;
-		var festival = this.props.params.festival;
-		var query = festival.split('-').join('%20');
-
-		var festivalData,
-			festivalUrl = constants.API_ROOT + 'festival/search/' + query;
-
-		$.ajax({
-			url: festivalUrl,
-			type: 'get',
+	getFestivalData(id) {
+		// test 452 Ultra 2015
+		api.get(`events/id/${id}`).then(res => {
+			var f = res.events_id
+			this.setState({
+				festival: f.event,
+				setCount: f.set_count,
+				sets: f.sets,
+				imageURL: f.banner_image.imageURL
+			});
+		}).then(() => {
+			this.setState({ loaded: true })
 		})
-		.done(function(response) {
-			festivalData = response.payload.festival;
-
-			push({
-				type: 'SHALLOW_MERGE',
-				data: {
-					detailId: festivalData.id,
-					detailData: festivalData
-				}
-			});
-
-			_this.setState({
-				loaded: true
-			});
-		});
 	},
 
-	render: function() {
-		var appState = this.props.appState;
-		var push = this.props.push;
-
-		var loginStatus = this.props.appState.get('isUserLoggedIn');
-		var user = this.props.appState.get('user');
-		var data = this.props.appState.get('detailData');
-
-		var setText = data.set_count > 1 ? ' sets' : ' set';
+	render() {
+		var setText = this.state.setCount != 1 ? 'sets' : 'set';
+		var festivalInfo = `${this.state.setCount} ${setText}`;
 
 		var detailInfo = {
-			appState: appState,
-			push: push,
-			title: data.event,
-			buttonText: 'Shuffle',
-			imageURL: data.main_imageURL,
-			info: data.set_count + setText
-		};
-
-		var setProps = {
-			containerClass: 'flex-row flex',
-			sets: data.sets,
-			push: push,
-			loginStatus: loginStatus,
-			user: user
+			title: this.state.festival,
+			info: festivalInfo,
+			imageURL: this.state.imageURL,
+			sets: R.pluck('id', this.state.sets)
 		};
 
 		return (
 			<Loader loaded={this.state.loaded}>
 				<div id='detail' className='view detail-page'>
 					<DetailImageContainer {...detailInfo}/>
-					<div className="flex-row links-container">
+					<div className='flex-row links-container'>
 						<div className='center flex-fixed'>
 							SETS
 						</div>
 					</div>
-					<SetContainer {...setProps} />
+					<SetContainer sets={this.state.sets} />
 				</div>
 			</Loader>
 		);
 	}
 });
 
-module.exports = FestivalDetail;
+export default FestivalDetail;
