@@ -1,31 +1,17 @@
 import React from 'react';
 import { Link } from 'react-router';
 
-var HomeSidebar = React.createClass({
-
-	getDefaultProps: function() {
-		return {
-			user: {}
-		};
+const HomeSidebar = React.createClass({
+	contextTypes: {
+		user: React.PropTypes.object
 	},
-
-	getProfilePic () {
-		
-	},
-
-	render: function() {
-		var name = this.props.user.first_name + ' ' + this.props.user.last_name;
-
+	render() {
 		return (
 			<div id='HomeSidebar'className='flex-column flex-fixed'>
-				<div className='flex flex-column overlay-container user-background hidden'>
-					<img className='center' src='' />
-					<div className='center'>{name}</div>
-				</div>
 				<div className='flex-3x flex-column user-nav'>
 					<div className='flex flex-row nav-header'>
 						<i className='fa fa-2x fa-user center'/>
-						<div className='center'>{name}</div>
+						<div className='center'>{this.context.user.first_name + ' ' + this.context.user.last_name}</div>
 					</div>
 					<Link className='nav-button click flex flex-row' to='/user' activeClassName='active' onlyActiveOnIndex={true}>
 					   <div>Favorites</div>
@@ -34,14 +20,13 @@ var HomeSidebar = React.createClass({
 					   <div>New Sets</div>
 					</Link>
 					<Link className='nav-button click flex flex-row' to='/user/events' activeClassName='active'>
-					   <div>New Events</div>
+						<div>New Events</div>
 					</Link>
 					<div className='buffer-3x'/>
 				</div>
 			</div>
 		);
 	}
-
 });
 
-module.exports = HomeSidebar;
+export default HomeSidebar;
