@@ -1,17 +1,14 @@
 import React from 'react';
-import R from 'ramda';
 import constants from '../constants/constants';
-import {History} from 'react-router';
 
 import LandingHome from './LandingHome';
 import LandingApp from './LandingApp';
 import LandingBeacon from './LandingBeacon';
 import Footer from './Footer';
 
-var LandingView = React.createClass({
+const LandingView = React.createClass({
 
 	displayName: 'Home Page',
-	mixins: [History],
 
 	render() {
 		var playerHidden = this.props.appState.get('playerHidden');
@@ -33,4 +30,4 @@ var LandingView = React.createClass({
 
 });
 
-module.exports = LandingView;
+export default LandingView;
