@@ -2,31 +2,29 @@ import React from 'react';
 import {Link} from 'react-router';
 
 var NavMenu = React.createClass({
-
-	getDefaultProps: function() {
+	getDefaultProps() {
 		return {
 			navItems: []
 		};
 	},
-	render: function() {
-		var links = this.props.navItems.map(function(nav, index){
-
+	render() {
+		var links = this.props.navItems.map((nav, index) => {
 			return(
 				<Link 
 					className='click flex flex-row' 
-					to={nav.link} 
+					to={nav.link}
 					key={index}
 					onlyActiveOnIndex={nav.onlyActiveOnIndex} 
 					activeClassName='active'>
 						<i className={nav.icon}/>
 						<div>{nav.text}</div>
 				</Link>
-			); 
+			);
 		});
 		
 		return (
 			<nav id='NavMenu' className='flex-column flex-fixed'>
-				<div className='set-flex flex'>
+				<div className='flex-container flex'>
 					<div className='nav-header'>BROWSE</div>
 				</div>
 				{links}
@@ -34,7 +32,6 @@ var NavMenu = React.createClass({
 			</nav>
 		);
 	}
-
 });
 
-module.exports = NavMenu;
+export default NavMenu;
