@@ -1,11 +1,14 @@
 import React from 'react';
-import {History} from 'react-router';
-import constants from '../constants/constants';
+import {Link} from 'react-router';
+import {API_ROOT} from '../constants/constants';
+import history from '../services/history';
 import _ from 'underscore';
 
 var SearchBar = React.createClass({
 
-	mixins: [History],
+	contextTypes: {
+		push: React.PropTypes.func
+	},
 
 	handleKeypress(e) {
 		var query = document.getElementById('search').value;
@@ -15,41 +18,35 @@ var SearchBar = React.createClass({
 	},
 
 	search(query) {
-		var self = this;
-		var push = this.props.push;
-		var activeSearchAjax = null;
-		var results, 
-			searchUrl = constants.API_ROOT + 'search/' + query;
+		var searchUrl = `${API_ROOT}search/${query}`;
 
 		$.ajax({
 			url: searchUrl,
 			type: 'get'
-		})
-		.done(function(response) {
-			results = response.payload.search;
-			var sets = results.sets;
-			var events = results.upcomingEvents;
-			var tracks = results.tracks;
-
-			push({
-				type: 'SHALLOW_MERGE',
-				data: {
-					searchResults: {
-						sets: sets,
-						upcomingEvents: events,
-						tracks: tracks
+		}).done(res => {
+			if(res.status === 'success') {
+				var {artists, sets, events, tracks} = res.payload.search;
+				this.context.push({
+					type: 'SHALLOW_MERGE',
+					data: {
+						searchResults: {
+							sets: sets,
+							upcomingEvents: events,
+							tracks: tracks,
+							artists: artists
+						}
 					}
-				}
-			});
+				});
 
-			self.history.pushState(null, '/search');
+				history.pushState(null, '/search');
+			}
 		});
 	},
 	
 	render() {
 		return (
 			<div className='center flex flex-row'>
-				<i className='nav-button fa fa-search center click'/>
+				<Link className='nav-button fa fa-search center click' to='/search' />
 				<input id='search' 
 					className='flex'
 					placeholder='search' 
@@ -61,4 +58,4 @@ var SearchBar = React.createClass({
 });
 
 
-module.exports = SearchBar;
+export default SearchBar;
