@@ -1,77 +1,79 @@
 import React from 'react';
-import SetTile from './SetTile';
-import favoriteSet from '../services/favoriteSet';
+import R from 'ramda';
+import {checkIfFavorited} from '../services/favoriteSet';
 
-var SetContainer = React.createClass({
+import SetTile from './SetTile';
+
+const SetContainer = React.createClass({
+	contextTypes: {
+		push: React.PropTypes.func,
+		user: React.PropTypes.object,
+		loginStatus: React.PropTypes.bool,
+		favoriteSetIds: React.PropTypes.array
+	},
+
+	checkIfFavorited(id, favorites) {
+		if(this.context.loginStatus) {
+			return R.contains(id, favorites);
+		} else {
+			return false
+		}
+	},
 
 	getDefaultProps() {
 		return {
-			containerClass: 'flex-row tile-container',
-			sets: [],
-			user: {
-				favorite_set_ids: []
-			}
+			className: 'flex-row tile-container scrollable',
+			sets: []
 		};
 	},
 
-	shouldComponentUpdate(nextProps, nextState) {
-		var oldFav = this.props.user.favorite_set_ids;
-		var newFav = nextProps.user.favorite_set_ids;
+	shouldComponentUpdate(nextProps, nextState, nextContext) {
+		var oldFav = this.context.favoriteSetIds
+		var newFav = nextContext.favoriteSetIds
 
-		if(nextProps.sets != this.props.sets) {
-			return true;
-		} else if(newFav != oldFav) {
-			return true;
-		} else {
-			return false;
+		switch(true) {
+			case nextProps.sets != this.props.sets:
+				return true;
+			case newFav != oldFav:
+				return true;
+			default:
+				return false;
+				break;
 		}
 	},
 
 	render() {
-		var data = this.props.sets;
-		var push = this.props.push;
-		var self = this;
+		var tiles = this.props.sets.map((set, index) => {
+			// check if each set is favorited
+			// var favorited = this.checkIfFavorited(set.id, this.context.favoriteSetIds)
+			var favorited = checkIfFavorited(this.context.loginStatus, set.id, this.context.favoriteSetIds)
 
-		var tiles = data.map(function(set, index) {
-			if(self.props.loginStatus) {
-				var favorited = favoriteSet.checkFavorite(set.id, self.props.user.favorite_set_ids);
+			// TODO doesn't work for user sets
+			if(set.episode != undefined && R.keys(set.episode).length != 0) {
+				var setName = `${set.event.event} - ${set.episode.episode}`;
 			} else {
-				var favorited = false;
+				var setName = set.event.event;
 			}
-
-			if(set.episode != null && set.episode.length > 0) {
-				var setName = set.event+' - '+set.episode	;
-			} else {
-				var setName = set.event;
-			}
-
-			var props = {
+			
+			return React.createElement(SetTile, {
 				key: index,
 				id: set.id,
-				artist_id: set.artist_id,
-				event_id: set.event_id,
-				is_radiomix: set.is_radiomix,
-				set_length: set.set_length,
-				push: push,
-				artist: set.artist,
-				event: set.event,
+				eventId: set.event_id,
+				isRadiomix: set.event.is_radiomix,
+				setLength: set.set_length,
+				artist: set.artists[0].artist,
+				event: set.event.event,
 				setName: setName,
-				artistimageURL: set.artistimageURL,
+				artistImage: set.artists[0].icon_image.imageURL_small,
 				popularity: set.popularity,
-				songURL: set.songURL,
-				main_eventimageURL: set.main_eventimageURL,
-				timePosition: set.timePosition,
-				user: self.props.user,
-				loginStatus: self.props.loginStatus,
+				songUrl: set.songURL,
+				bannerImage: set.event.banner_image.imageURL,
 				favorited: favorited
-			};
-
-			return(<SetTile {...props} />);
+			})
 		});
 
-
 		return (
-			<div className={this.props.containerClass}>
+			<div className={this.props.className}>
 				{tiles}
 			</div>
 		);
@@ -79,4 +81,4 @@ var SetContainer = React.createClass({
 
 });
 
-module.exports = SetContainer;
+export default SetContainer;
