@@ -1,8 +1,9 @@
 import React from 'react';
-import constants from '../constants/constants';
-import mixpanelService from './mixpanelService.js';
+import R from 'ramda';
+import api from './api';
+import mixpanelService from './mixpanelService';
 
-function startFacebookSDK(push) {
+export function startFacebookSDK(push) {
 	window.fbAsyncInit = function() {
 		FB.init({
 			appId      : '648288801959503',
@@ -11,7 +12,6 @@ function startFacebookSDK(push) {
 			xfbml      : true,  // parse social plugins on this page
 			version    : 'v2.1' // use version 2.1
 		});
-
 			// Now that we've initialized the JavaScript SDK, we call
 			// FB.getLoginStatus().  This function gets the state of the
 			// person visiting this page and can return one of three states to
@@ -39,21 +39,17 @@ function startFacebookSDK(push) {
 }
 
 function statusChangeCallback(response, push) {
-	console.log('statusChangeCallback');
-	console.log(response);
 	if (response.status === 'connected') {
-		// Logged into your app and Facebook.
-		registerFacebookUser(response.authResponse, push)
+		// Logged into setmine and Facebook.
+		registerFacebookUser(response.authResponse.accessToken, push)
 	} else if (response.status === 'not_authorized') {
-		// The person is logged into Facebook, but not your app.
-			
+		console.log('Logged into Facebook, but you need to authorize this app');
 	} else {
-		// The person is not logged into Facebook, so we're not sure if
-		// they are logged into this app or not.
-		console.debug("Not logged into Facebook");
+		console.debug('Not logged into Facebook');
 	}
 }
 
+// check if user is logged in
 function checkLoginState(push) {
 	FB.getLoginStatus(function(response) {
 		statusChangeCallback(response, push);
@@ -61,40 +57,38 @@ function checkLoginState(push) {
 }
 
 function registerFacebookUser(auth, push) {
-	$.ajax({
-		type: 'POST',
-		url: constants.API_ROOT + 'user/facebookRegister',
-		data: {
-			userData: {
-				FB_TOKEN: auth
-			}
-		}
-	}).done(function(res) {
+	api.post('setmineuser/login/facebook', {
+		facebook_token: auth
+	}).then(res => {
+		console.log('Successfully logged in to Setmine')
+		var user = res.setmineuser_login_facebook
+
+		// create array of users favorite setIds
+		var favoriteSetIds = R.pluck('id', user.favorite_sets)
+
+		// store setmine user in appState
 		push({
 			type: 'SHALLOW_MERGE',
 			data: {
 				isUserLoggedIn: true,
-				user: res.payload.user
+				user: user,
+				favoriteSetIds: favoriteSetIds
 			}
-		});
+		})
 
 		//track user after logging in for the first time
-		mixpanel.identify(res.payload.user.facebook_id);
+		mixpanel.identify(user.facebook_id);
 		mixpanel.people.set_once({
-			"First Name": res.payload.user.first_name,
-			"Last Name": res.payload.user.last_name,
-			"$email": res.payload.user.username,
-			"fb_id": res.payload.user.facebook_id,
+			"First Name": user.first_name,
+			"Last Name": user.last_name,
+			"$email": user.username,
+			"fb_id": user.facebook_id,
 			"date_tracked": new Date()
 		});
-	});
+	})
 }
 
-function login(push) {
+// starts login process
+export function login(push) {
 	FB.login(checkLoginState(push));
-}
-
-module.exports = {
-	login: login,
-	startFacebookSDK: startFacebookSDK
 }
