@@ -1,12 +1,7 @@
-
-var splice = {
-	 bigArray: function(array, cutoff) {
-	    if(array.length > cutoff) {
-			return array.splice(cutoff, array.length-cutoff);
-	    } else {
-	    	return array;
-	    }
+export function bigArray(array, cutoff) {
+	if(array.length > cutoff) {
+		return array.splice(cutoff, array.length-cutoff);
+	} else {
+		return array;
 	}
-};
-
-module.exports = splice;
+}
